\docType{package}
\name{ConnMatTools}
\alias{ConnMatTools}
\alias{ConnMatTools-package}
\title{Tools for working with connectivity matrices}
\description{
Tools for working with connectivity matrices
}
\details{
ConnMatTools collects several different methods for
analyzing and working with connectivity matrices in R.
Though primarily oriented towards marine larval dispersal,
many of the methods are general and useful for terrestrial
systems as well.

\tabular{ll}{ Package: \tab ConnMatTools\cr Type: \tab
Package\cr Version: \tab 0.1.4\cr Date: \tab 2013-07-09\cr
License: \tab GPL (>= 2)\cr LazyLoad: \tab no\cr }
}
\examples{
\dontrun{optimalSplitConnMat(CM)}
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com} Marco
Andrello \email{marco.andrello@gmail.com}

Maintainer: David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\seealso{
See \code{\link{optimalSplitConnMat}}
}
\keyword{package}

