% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detecting_marked_individuals.R
\encoding{UTF-8}
\name{dmixfunc.norm}
\alias{dmixfunc.norm}
\alias{pmixfunc.norm}
\alias{qmixfunc.norm}
\title{Functions for examining the probability distribution for the fraction of
marked individuals}
\usage{
dmixfunc.norm(obs, d.unmarked, d.marked, N = .dmixfunc.N(obs), ...)

pmixfunc.norm(obs, d.unmarked, d.marked, N = .dmixfunc.N(obs), ...)

qmixfunc.norm(obs, d.unmarked, d.marked, N = .dmixfunc.N(obs), ...)
}
\arguments{
\item{obs}{Vector of observed score values for potentially marked individuals}

\item{d.unmarked}{A function representing the PDF of unmarked individuals.
Must be normalized so that it integrates to 1 for the function to work
properly.}

\item{d.marked}{A function representing the PDF of marked individuals.  Must
  be normalized so that it integrates to 1 for the function to work properly.}

\item{N}{number of steps in fixed-step, trapezoidal integration scheme used
to normalize probability distributions.  Defaults to \code{2*length(obs)}
so long as that number is comprised between \code{100} and \code{5000}.}

\item{\dots}{Additional arguments for the \code{\link{approxfun}} function.}
}
\value{
A function that takes one argument (the fraction of marked
  individuals for \code{dmixfunc.norm} and \code{pmixfunc.norm}; the quantile
  for \code{qmixfunc.norm}) and returns the probability density, cumulative
  probability distribution or score value, respectively.
}
\description{
These functions return functions that calculates the probability density
function (\code{dmixfunc.norm}), the probability distribution function (aka
the cumulative distribution function; \code{pmixfunc.norm}) and the quantile
function (\code{qmixfunc.norm}) given a set of observed score values and
distributions for unmarked and marked individuals.
}
\details{
The normalization of the probability distribution is carried out using a
simple, fixed-step trapezoidal integration scheme.  The number of steps
between marked fractions 0 and 1 defaults to \code{2*length(obs)} so long as
that number is comprised between \code{100} and \code{5000}.
}
\section{Functions}{
\itemize{
\item \code{dmixfunc.norm}: Returns a function that is PDF for fraction of
marked individuals

\item \code{pmixfunc.norm}: Returns a function that is cumulative probability
distribution for fraction of marked individuals

\item \code{qmixfunc.norm}: Returns a function that is quantile function for
fraction of marked individuals
}}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}

