% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{gamma.mean.sd.shape.scale}
\alias{gamma.mean.sd.shape.scale}
\title{Gamma distribution shape and scale parameters from mean and standard
deviation, or vice-versa}
\usage{
\method{gamma}{mean.sd.shape.scale}(...)
}
\arguments{
\item{mean}{Mean of the gamma distribution}

\item{sd}{Standard deviation of the gamma distribution}

\item{shape}{Shape parameter of the gamma distribution}

\item{scale}{Scale parameter of the gamma distribution}
}
\value{
A list with \code{mean}, \code{sd}, \code{shape} and \code{scale}
  parameters of the corresponding gamma distribution.
}
\description{
Calculates shape and scale parameters for a gamma distribution from the mean
and standard deviation of the distribution, or vice-versa.  One supplies
either \code{mean} and \code{sd} or \code{shape} and \code{scale} and the
function returns a list with all four parameter values.
}
\examples{
mn <- 1
sd <- 0.4
l <- gamma.mean.sd.shape.scale(mean=mn,sd=sd)
x <- seq(0,2,length.out=50)

# Plot gamma and normal distributions - for sd << mean, the two should be very close
plot(x,dgamma(x,l$shape,scale=l$scale),
     main="Normal versus Gamma distributions",type="l")
lines(x,dnorm(x,l$mean,l$sd),col="red")
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}

