% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detecting_marked_individuals.R
\encoding{UTF-8}
\name{log.prob}
\alias{log.prob}
\title{Logarithm of probability for a set of observed score values given a PDF}
\usage{
\method{log}{prob}(p, obs, dfunc)
}
\arguments{
\item{p}{Parameter in the PDF distribution.  This should be a single scalar
value, presumably the fraction of marked individuals in the population, but
the function is generic.}

\item{obs}{A vector of score values}

\item{dfunc}{A function representing the PDF of scores.  Should be normalized
  so that it integrates to 1 for the function to return a true
  log-probability.}
}
\value{
The log-probability.
}
\description{
This function returns log-probability for a set of observed score values
given a PDF for the distribution of scores.
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}

