% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detecting_marked_individuals.R
\encoding{UTF-8}
\name{optim.rel.conn.dists}
\alias{optim.rel.conn.dists}
\title{Calculate the optimal value for relative connectivity given two distributions
for scores for marked and unmarked individuals}
\usage{
optim.rel.conn.dists(obs, d.unmarked, d.marked, p = 1, phi0 = 0.5,
  method = "Brent", lower = 0, upper = 1, ...)
}
\arguments{
\item{obs}{Vector of observed score values for potentially marked individuals}

\item{d.unmarked}{A function representing the PDF of unmarked individuals.
Must be normalized so that it integrates to 1 for the function to work
properly.}

\item{d.marked}{A function representing the PDF of marked individuals.  Must
  be normalized so that it integrates to 1 for the function to work properly.}

\item{p}{Fraction of individuals (i.e., eggs) marked in the source population}

\item{phi0}{Initial value for \eqn{\phi}, the fraction of settlers at the
destination population that originated at the source population, for
\code{\link{optim}} function. Defaults to 0.5.}

\item{method}{Method variable for \code{\link{optim}} function. Defaults to
\code{"Brent"}.}

\item{lower}{Lower limit for search for fraction of marked individuals.
Defaults to 0.}

\item{upper}{Upper limit for search for fraction of marked individuals.
Defaults to 1.}

\item{\dots}{Additional arguments for the \code{\link{optim}} function.}
}
\value{
A list with results of optimization. Optimal fraction of marked
  individuals is in \code{par} field. Negative log-likelihood is in the
  \code{value} field. See \code{\link{optim}} for more details.
}
\description{
This function calculates the value for relative connectivity that best fits a
set of observed score values, a pair of distributions for marked and unmarked
individuals and an estimate of the fraction of eggs marked in the source
population, \code{p}.
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}
\seealso{
See also \code{\link{d.relative.connectivity}}
}

