% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detecting_marked_individuals.R
\encoding{UTF-8}
\name{optimal.fraction.from.distributions}
\alias{optimal.fraction.from.distributions}
\title{Calculate the optimal fraction of marked individuals}
\usage{
optimal.fraction.from.distributions(obs, d.unmarked, d.marked, par = 0.5,
  method = "Brent", lower = 0, upper = 1, ...)
}
\arguments{
\item{obs}{Vector of observed score values for potentially marked individuals}

\item{d.unmarked}{A function representing the PDF of unmarked individuals.
Must be normalized so that it integrates to 1 for the function to work
properly.}

\item{d.marked}{A function representing the PDF of marked individuals.  Must
  be normalized so that it integrates to 1 for the function to work properly.}

\item{par}{Initial fraction of marked individuals for \code{\link{optim}}
function. Defaults to 0.5.}

\item{method}{Method variable for \code{\link{optim}} function. Defaults to
\code{"Brent"}.}

\item{lower}{Lower limit for search for fraction of marked individuals.
Defaults to 0.}

\item{upper}{Upper limit for search for fraction of marked individuals.
Defaults to 1.}

\item{\dots}{Additional arguments for the \code{\link{optim}} function.}
}
\value{
A list with results of optimization. Optimal fraction of marked
  individuals is in \code{par} field. See \code{\link{optim}} for details.
}
\description{
This function calculates the fraction of marked individuals that best fits a
set of observed score values and a pair of distributions for marked and
unmarked individuals.
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}

