% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimating_marked_egg_fraction.R
\encoding{UTF-8}
\name{r.marked.egg.fraction}
\alias{r.marked.egg.fraction}
\title{Estimates of fraction of eggs marked accounting for variability in
reproductive output}
\usage{
r.marked.egg.fraction(n, n.females, n.marked.females = round(n.females *
  p.marked.females), mean.female = 1, cv.female, dual = FALSE,
  male.uncert = FALSE, n.males = n.females,
  n.marked.males = tryCatch(round(n.males * p.marked.males), error =
  function(e) n.marked.females), mean.male = mean.female,
  cv.male = cv.female, p.marked.females, p.marked.males = p.marked.females)
}
\description{
Estimates of fraction of eggs marked accounting for variability in
reproductive output
}
\examples{
library(ConnMatTools)
data(damselfish.lods)

# Histograms of simulated LODs
l <- seq(-1,30,0.5)
h.in <- hist(damselfish.lods$in.group,breaks=l)
h.out <- hist(damselfish.lods$out.group,breaks=l)

# PDFs for marked and unmarked individuals based on simulations
d.marked <- stepfun.hist(h.in)
d.unmarked <- stepfun.hist(h.out)

# Fraction of adults genotyped at source site
p.adults <- 0.25

# prior.shape1=1 # Uniform prior
prior.shape1=0.5 # Jeffreys prior

# Fraction of eggs from one or more genotyped parents
p <- dual.mark.transmission(p.adults)$p

# PDF for relative connectivity
D <- d.rel.conn.dists.func(damselfish.lods$real.children,
                           d.unmarked,d.marked,p,
                           prior.shape1=prior.shape1)

# Estimate most probable value for relative connectivity
phi.mx <- optim.rel.conn.dists(damselfish.lods$real.children,
                                    d.unmarked,d.marked,p)$phi

# Estimate 95\% confidence interval for relative connectivity
Q <- q.rel.conn.dists.func(damselfish.lods$real.children,
                           d.unmarked,d.marked,p,
                           prior.shape1=prior.shape1)

# Plot it up
phi <- seq(0,1,0.001)
plot(phi,D(phi),type="l",
     xlim=c(0,0.1),
     main="PDF for relative connectivity",
     xlab=expression(phi),
     ylab="Probability density")

abline(v=phi.mx,col="green",lty="dashed")
abline(v=Q(c(0.025,0.975)),col="red",lty="dashed")
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}
\seealso{
Other connectivity.estimation: \code{\link{d.rel.conn.beta.prior}},
  \code{\link{p.rel.conn.beta.prior}},
  \code{\link{q.rel.conn.beta.prior}},
  \code{\link{q.rel.conn.beta.prior.func}};
  \code{\link{d.rel.conn.dists.func}},
  \code{\link{p.rel.conn.dists.func}},
  \code{\link{q.rel.conn.dists.func}};
  \code{\link{d.rel.conn.finite.settlement}};
  \code{\link{d.rel.conn.multiple}},
  \code{\link{p.rel.conn.multiple}},
  \code{\link{q.rel.conn.multiple}},
  \code{\link{q.rel.conn.multiple.func}};
  \code{\link{d.rel.conn.unif.prior}},
  \code{\link{p.rel.conn.unif.prior}},
  \code{\link{q.rel.conn.unif.prior}};
  \code{\link{dual.mark.transmission}};
  \code{\link{optim.rel.conn.dists}}
}

