% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/detecting_marked_individuals.R
\encoding{UTF-8}
\name{stepfun.hist}
\alias{stepfun.hist}
\title{Create a probability density step function from a histogram object}
\usage{
stepfun.hist(h, ..., normalize = TRUE)
}
\arguments{
\item{h}{an object of type \code{histogram}}

\item{normalize}{Boolean indicating whether or not to normalize the output
  stepfun so that it integrates to 1. Defaults to \code{TRUE}.  If
  \code{FALSE}, then the function will integrate to \code{sum(h$counts)}}

\item{\dots}{Additional arguments for the default \code{\link{stepfun}}
function.}
}
\value{
A function of class \code{stepfun}.  The height of the steps will be
  divided by the distance between breaks and possibly the
}
\description{
This function creates a step function from the bars in a \code{histogram}
object. By default, the step function will be normalized so that it
integrates to 1.
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}

