% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{chile.loco}
\alias{chile.loco}
\title{Connectivity matrix for loco (Concholepas concholepas) from Chile}
\description{
Sample connectivity matrix representing potential larval dispersal of loco
(Concholepas concholepas) from Chile.  The matrix is for 89 sites along the
coast of Chile and is derived from a theoretical larval transport model.
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\keyword{data}

