% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\encoding{UTF-8}
\name{chile.loco}
\alias{chile.loco}
\title{Connectivity matrix for loco (Concholepas concholepas) from Chile}
\format{A square 89x89 matrix with real, positive elements.}
\description{
Sample connectivity matrix representing potential larval dispersal of loco 
(Concholepas concholepas) from Chile.  The matrix is for 89 sites along the 
coast of Chile and is derived from a theoretical larval transport model.
}
\references{
Garavelli L, Kaplan DM, Colas F, Stotz W, Yannicelli B, Lett C
  (2014) Identifying appropriate spatial scales for marine conservation and
  management using a larval dispersal model: The case of Concholepas
  concholepas (loco) in Chile. Progress in Oceanography 124:42-53.
  \href{http://dx.doi.org/10.1016/j.pocean.2014.03.011}{doi:10.1016/j.pocean.2014.03.011}
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\keyword{data}
