% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connectivity_estimation.distributions.R
\encoding{UTF-8}
\name{d.rel.conn.dists.func}
\alias{d.rel.conn.dists.func}
\alias{p.rel.conn.dists.func}
\alias{q.rel.conn.dists.func}
\title{Functions for estimating the probability distribution for relative
connectivity given a pair of distributions for scores for marked and unmarked
individuals}
\usage{
d.rel.conn.dists.func(obs, d.unmarked, d.marked, p = 1,
  N = .d.rel.conn.dists.N(obs), ...)

p.rel.conn.dists.func(obs, d.unmarked, d.marked, p = 1,
  N = .d.rel.conn.dists.N(obs), ...)

q.rel.conn.dists.func(obs, d.unmarked, d.marked, p = 1,
  N = .d.rel.conn.dists.N(obs), ...)
}
\arguments{
\item{obs}{Vector of observed score values for potentially marked individuals}

\item{d.unmarked}{A function representing the PDF of unmarked individuals.
Must be normalized so that it integrates to 1 for the function to work
properly.}

\item{d.marked}{A function representing the PDF of marked individuals.  Must
  be normalized so that it integrates to 1 for the function to work properly.}

\item{p}{Fraction of individuals (i.e., eggs) marked in the source
population. Defaults to 1.}

\item{N}{number of steps in fixed-step, trapezoidal integration scheme used
to normalize probability distributions.  Defaults to \code{2*length(obs)}
so long as that number is comprised between \code{100} and \code{5000}.}

\item{\dots}{Additional arguments for the \code{\link{approxfun}} function.}
}
\value{
A function that takes one argument (the relative connectivity for
  \code{d.rel.conn.dists.func} and \code{p.rel.conn.dists.func}; the quantile
  for \code{q.rel.conn.dists.func}) and returns the probability density,
  cumulative probability distribution or score value, respectively. The
  returned function accepts both vector and scalar input values.
}
\description{
These functions return functions that calculate the probability density
function (\code{d.rel.conn.dists.func}), the probability distribution
function (aka the cumulative distribution function;
\code{p.rel.conn.dists.func}) and the quantile function
(\code{q.rel.conn.dists.func}) for relative connectivity given a set of
observed score values, distributions for unmarked and marked individuals, and
an estimate of the fraction of all eggs marked at the source site, \code{p}.
}
\details{
The normalization of the probability distribution is carried out using a
simple, fixed-step trapezoidal integration scheme.  The number of steps
between relative connectivity values of 0 and 1 defaults to
\code{2*length(obs)} so long as that number is comprised between \code{100}
and \code{5000}.
}
\section{Functions}{
\itemize{
\item \code{d.rel.conn.dists.func}: Returns a function that is PDF for relative
connectivity

\item \code{p.rel.conn.dists.func}: Returns a function that is cumulative
probability distribution for relative connectivity

\item \code{q.rel.conn.dists.func}: Returns a function that is quantile
function for relative connectivity
}}
\examples{
library(ConnMatTools)
data(damselfish.lods)

# Histograms of simulated LODs
l <- seq(-1,30,0.5)
h.in <- hist(damselfish.lods$in.group,breaks=l)
h.out <- hist(damselfish.lods$out.group,breaks=l)

# PDFs for marked and unmarked individuals based on simulations
d.marked <- stepfun.hist(h.in)
d.unmarked <- stepfun.hist(h.out)

# Fraction of adults genotyped at source site
p.adults <- 0.25

# Fraction of eggs from one or more genotyped parents
p <- dual.mark.transmission(p.adults)$p

# PDF for relative connectivity
D <- d.rel.conn.dists.func(damselfish.lods$real.children,
                           d.unmarked,d.marked,p)

# Estimate most probable value for relative connectivity
phi.mx <- optim.rel.conn.dists(damselfish.lods$real.children,
                                    d.unmarked,d.marked,p)$phi

# Estimate 95\% confidence interval for relative connectivity
Q <- q.rel.conn.dists.func(damselfish.lods$real.children,
                           d.unmarked,d.marked,p)

# Plot it up
phi <- seq(0,1,0.01)
plot(phi,D(phi),type="l",
     xlim=c(0,0.1),
     main="PDF for relative connectivity",
     xlab=expression(phi),
     ylab="Probability density")

abline(v=phi.mx,col="green",lty="dashed")
abline(v=Q(c(0.025,0.975)),col="red",lty="dashed")
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}
\seealso{
See also \code{\link{d.relative.connectivity}}
}

