% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connectivity_estimation.finite_settlement.R
\encoding{UTF-8}
\name{d.rel.conn.finite.settlement}
\alias{d.rel.conn.finite.settlement}
\title{Frequency of the number of marked settlers for finite total settlement}
\usage{
d.rel.conn.finite.settlement(phi, p, n.obs, n.settlers, n.bootstraps = 1000)
}
\arguments{
\item{phi}{Vector of fractions of settlers at the destination population that
originated at the source population}

\item{p}{Fraction of individuals (i.e., eggs) marked in the source population}

\item{n.obs}{Total number of settlers collected}

\item{n.settlers}{Total number of settlers at the destination site from which
the \code{n.obs} (\code{<=n.settlers}) settlers are collected}

\item{n.bootstraps}{Number of simulated samples to generate to calculate
  relative frequencies. Defaults to \code{1000}.}
}
\value{
A matrix with \code{n.obs+1} rows and \code{length(phi)} columns, the
  element of which are counts indicating the number of times among the
  \code{n.bootstraps} samples that \code{k} marked settlers are found among
  the \code{n.obs} settlers collected from the total settlement pool of
  \code{n.settlers} individuals. Results for \code{k} marked settlers are in
  row \code{k+1} of the matrix, the first row being used for \code{k=0}.  The
  row names of the matrix reflect the value of \code{k}.
}
\description{
This function calculates the relative frequency with which a number of marked
settlers would be observed as a function of relative connectivity, sample
size and the total number of settlers to a site. Probable frequency of
observation is calculated by randomly drawing \code{n.bootstraps} artificial
samples and calculating for each the number of would-be marked settlers for
all values of \code{phi}, the relative connectivity value.
}
\examples{
library(ConnMatTools)

k <- 10 # Number of marked settlers among sample
n.obs <- 87 # Number of settlers in sample
n.settlers <- 100 # Total size of settler pool

p <- 0.4 # Fraction of eggs that was marked
phi <- seq(0,1,length.out=101) # Values for relative connectivity

drc <- d.relative.connectivity(phi,p,k,n.obs)
prc <- p.relative.connectivity(phi,p,k,n.obs)
qrc <- q.relative.connectivity(c(0.025,0.975),p,k,n.obs) # 95\% confidence interval

n.bootstraps <- 1000 # Increase to improve accuracy of brute force estimations

# Test with brute force finite settlement function
dbf <- d.rel.conn.finite.settlement(phi,p,n.obs,100*n.obs,
                                    n.bootstraps=n.bootstraps)
dbf <- dbf[as.character(k),]

# Normalize and calculate cumulative probability function
pbf <- cumsum(c(0,(dbf[1:length(dbf)-1]+dbf[2:length(dbf)])/2*diff(phi)))
dbf <- dbf / max(pbf)
pbf <- pbf / max(pbf)

# Approximate quantiles
qbf <- approx(pbf,phi,c(0.025,0.975))$y

# Finite recruitment
dfr <- d.rel.conn.finite.settlement(phi,p,n.obs,n.settlers,
                                    n.bootstraps=n.bootstraps)
dfr <- dfr[as.character(k),]

# Normalize and calculate cumulative probability function
pfr <- cumsum(c(0,(dfr[1:length(dfr)-1]+dfr[2:length(dfr)])/2*diff(phi)))
dfr <- dfr / max(pfr)
pfr <- pfr / max(pfr)

# Approximate quantiles
qfr <- approx(pfr,phi,c(0.025,0.975))$y

# Make a plot of different distributions
plot(phi,drc,type="l",main="Probability of relative connectivity values",
     xlab=expression(phi),ylab="Probability density")
lines(phi,prc,col="blue")
lines(phi,dbf,col="black",lty="dashed")
lines(phi,dfr,col="red",lty="dashed")
abline(v=qrc,col="black")
abline(v=qbf,col="black",lty="dashed")
abline(v=qfr,col="red",lty="dashed")
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}
\seealso{
See also \code{\link{d.relative.connectivity}},
  \code{\link{d.rel.conn.multiple}}
}

