% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connectivity_estimation.R
\encoding{UTF-8}
\name{d.rel.conn.unif.prior}
\alias{d.rel.conn.unif.prior}
\alias{p.rel.conn.unif.prior}
\alias{q.rel.conn.unif.prior}
\title{Estimate the probability distribution of relative connectivity values
assuming a uniform prior distribution}
\usage{
d.rel.conn.unif.prior(phi, p, k, n, log = FALSE, ...)

p.rel.conn.unif.prior(phi, p, k, n, log = FALSE, ...)

q.rel.conn.unif.prior(q, p, k, n, log = FALSE, ...)
}
\arguments{
\item{phi}{Vector of fractions of individuals (i.e., eggs) from the source
population settling at the destination population}

\item{p}{Fraction of individuals (i.e., eggs) marked in the source population}

\item{k}{Number of marked settlers found in sample}

\item{n}{Total number of settlers collected}

\item{log}{If \code{TRUE}, returns natural logarithm of probabilities, except
for \code{\link{q.rel.conn.unif.prior}}, which expects log of probabilities
as inputs}

\item{q}{Vector of quantiles}

\item{\dots}{Extra arguments to Beta distribution functions.  See
  \code{\link{dbeta}} for details.  For expert use only.}
}
\value{
Vector of probabilities or quantiles.
}
\description{
These functions calculate the probability density function
(\code{d.rel.conn.unif.prior}), the probability distribution function (aka
the cumulative distribution function; \code{p.rel.conn.unif.prior}) and the
quantile function (\code{q.rel.conn.unif.prior}) for the relative (to all
settlers at the destination site) connectivity value for larval transport
between a source and destination site given a known fraction of marked
individuals (i.e., eggs) in the source population.  A uniform prior is used
for the relative connectivity value.
}
\details{
Estimations of the probability distribution are derived from the Beta
distribution (see \code{\link{dbeta}}) and should be exact to great
precision.
}
\section{Functions}{
\itemize{
\item \code{d.rel.conn.unif.prior}: Returns the probability density for
relative connectivity between a pair of sites

\item \code{p.rel.conn.unif.prior}: Returns the cumulative probability
distribution for relative connectivity between a paire of sites

\item \code{q.rel.conn.unif.prior}: Estimates quantiles for the probability
distribution function for relative connectivity between a pair of sites
}}
\examples{
library(ConnMatTools)

k <- 10 # Number of marked settlers among sample
n.obs <- 87 # Number of settlers in sample
n.settlers <- 100 # Total size of settler pool

p <- 0.4 # Fraction of eggs that was marked
phi <- seq(0,1,length.out=101) # Values for relative connectivity

# Probability distribution assuming infinite settler pool and uniform prior
drc <- d.rel.conn.unif.prior(phi,p,k,n.obs)
prc <- p.rel.conn.unif.prior(phi,p,k,n.obs)
qrc <- q.rel.conn.unif.prior(c(0.025,0.975),p,k,n.obs) # 95\% confidence interval

n.bootstraps <- 1000 # Increase to improve accuracy of brute force estimations

# Test with brute force finite settlement function and (nearly) infinite settler pool
dbf <- d.rel.conn.finite.settlement(phi,p,n.obs,100*n.obs,
                                    n.bootstraps=n.bootstraps)
dbf <- dbf[as.character(k),]

# Normalize and calculate cumulative probability function
pbf <- cumsum(c(0,(dbf[1:length(dbf)-1]+dbf[2:length(dbf)])/2*diff(phi)))
dbf <- dbf / max(pbf)
pbf <- pbf / max(pbf)

# Approximate quantiles
qbf <- approx(pbf,phi,c(0.025,0.975))$y

# Finite settler pool
dfr <- d.rel.conn.finite.settlement(phi,p,n.obs,n.settlers,
                                    n.bootstraps=n.bootstraps)
dfr <- dfr[as.character(k),]

# Normalize and calculate cumulative probability function
pfr <- cumsum(c(0,(dfr[1:length(dfr)-1]+dfr[2:length(dfr)])/2*diff(phi)))
dfr <- dfr / max(pfr)
pfr <- pfr / max(pfr)

# Approximate quantiles
qfr <- approx(pfr,phi,c(0.025,0.975))$y

# Make a plot of different distributions
plot(phi,drc,type="l",main="Probability of relative connectivity values",
     xlab=expression(phi),ylab="Probability density")
lines(phi,prc,col="blue")
lines(phi,dbf,col="black",lty="dashed")
lines(phi,dfr,col="red",lty="dashed")
abline(v=qrc,col="black")
abline(v=qbf,col="black",lty="dashed")
abline(v=qfr,col="red",lty="dashed")
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}
\seealso{
Other connectivity.estimation: \code{\link{d.rel.conn.beta.prior}},
  \code{\link{p.rel.conn.beta.prior}},
  \code{\link{q.rel.conn.beta.prior}},
  \code{\link{q.rel.conn.beta.prior.func}};
  \code{\link{d.rel.conn.dists.func}},
  \code{\link{p.rel.conn.dists.func}},
  \code{\link{q.rel.conn.dists.func}};
  \code{\link{d.rel.conn.finite.settlement}};
  \code{\link{d.rel.conn.multiple}},
  \code{\link{p.rel.conn.multiple}},
  \code{\link{q.rel.conn.multiple}},
  \code{\link{q.rel.conn.multiple.func}};
  \code{\link{dual.mark.transmission}};
  \code{\link{optim.rel.conn.dists}}
}

