% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connectivity_estimation.R
\encoding{UTF-8}
\name{d.relative.connectivity}
\alias{d.relative.connectivity}
\alias{p.relative.connectivity}
\alias{q.relative.connectivity}
\alias{q.relative.connectivity.func}
\title{Functions for estimating the probability distribution of relative
connectivity values}
\usage{
d.relative.connectivity(phi, p, k, n)

p.relative.connectivity(phi, p, k, n)

q.relative.connectivity.func(p, k, n, N = 1000)

q.relative.connectivity(q, p, k, n, N = 1000)
}
\arguments{
\item{phi}{Vector of fractions of individuals (i.e., eggs) from the source
population settling at the destination population}

\item{p}{Fraction of individuals (i.e., eggs) marked in the source population}

\item{k}{Number of marked settlers found in sample}

\item{n}{Total number of settlers collected}

\item{N}{Number of points at which to estimate cumulative probability
  function for reverse approximation of quantile distribution. Defaults to
  \code{1000}.}

\item{q}{Vector of quantiles}
}
\value{
Vector of probabilities or quantiles.
}
\description{
These functions calculate the probability density function
(\code{d.relative.connectivity}), the probability distribution function (aka
the cumulative distribution function; \code{p.relative.connectivity}) and the
quantile function (\code{q.relative.connectivity}) for the relative (to all
settlers at the destination site) connectivity value for larval transport
between a source and destination site given a known fraction of marked
individuals (i.e., eggs) in the source population.
}
\details{
Estimations of the probability distribution are analytic, except that
quantile estimation is performed using \code{\link{approxfun}} to perform
reverse estimation based on the cumulative probability distribution function
estimated at a finite number of points.
}
\section{Functions}{
\itemize{
\item \code{d.relative.connectivity}: Returns the probability density for
relative connectivity between a paire of sites

\item \code{p.relative.connectivity}: Returns the cumulative probability
distribution for relative connectivity between a paire of sites

\item \code{q.relative.connectivity.func}: Returns a function to estimate quantiles
for the probability distribution function for relative connectivity between
a pair of sites

\item \code{q.relative.connectivity}: Estimates quantiles for the probability
distribution function for relative connectivity between a pair of sites
}}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}

