% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connectivity_estimation.distributions.R
\encoding{UTF-8}
\name{prob.marked}
\alias{prob.marked}
\title{Returns probability a set of observations correspond to marked individuals}
\usage{
prob.marked(obs, d.unmarked, d.marked, phi, p = 1)
}
\arguments{
\item{obs}{A vector of score values for a random sample of (marked and
unmarked) individuals from the population}

\item{d.unmarked}{A function representing the PDF of unmarked individuals.
Must be normalized so that it integrates to 1 for the function to work
properly.}

\item{d.marked}{A function representing the PDF of marked individuals.  Must
  be normalized so that it integrates to 1 for the function to work properly.}

\item{phi}{The fraction of settlers at the destination population that
originated at the source population}

\item{p}{Fraction of individuals (i.e., eggs) marked in the source
  population. Defaults to 1.}
}
\value{
A vector of the same size as \code{obs} containing the probability
  that each individual is marked
}
\description{
This function returns the probability each of a set of observations
corresponds to a marked individual given the distribution of scores for
unmarked and marked individuals and the fraction of individuals that are
marked.
}
\author{
David M. Kaplan \email{dmkaplan2000@gmail.com}
}
\references{
Kaplan et al. (submitted) Uncertainty in marine larval
  connectivity estimation
}

